+++ src/gridstack.js
@@ -18,6 +18,8 @@
     factory(window.jQuery, window);
   }
 })(function($, scope) {
+
+  // checks for obsolete method names
   var obsolete = function(f, oldName, newName) {
     var wrapper = function() {
       console.warn('gridstack.js: Function `' + oldName + '` is deprecated as of v0.2.5 and has been replaced ' +
@@ -29,9 +31,23 @@
     return wrapper;
   };
 
-  var obsoleteOpts = function(oldName, newName) {
-    console.warn('gridstack.js: Option `' + oldName + '` is deprecated as of v0.2.5 and has been replaced with `' +
-      newName + '`. It will be **completely** removed in v1.0.');
+  // checks for obsolete grid options 9can be used for any fields, but msg is about options)
+  var obsoleteOpts = function(opts, oldName, newName) {
+    if (opts[oldName] !== undefined) {
+      opts[newName] = opts[oldName];
+      console.warn('gridstack.js: Option `' + oldName + '` is deprecated as of v0.2.5 and has been replaced with `' +
+        newName + '`. It will be **completely** removed in v1.0.');
+    }
+  };
+
+  // checks for obsolete Jquery element attributes
+  var obsoleteAttr = function(el, oldName, newName) {
+    var oldAttr = el.attr(oldName);
+    if (oldAttr !== undefined) {
+      el.attr(newName, oldAttr);
+      console.warn('gridstack.js: attribute `' + oldName + '`=' + oldAttr + ' is deprecated on this object as of v0.5.2 and has been replaced with `' +
+        newName + '`. It will be **completely** removed in v1.0.');
+    }
   };
 
   var Utils = {
@@ -279,10 +295,10 @@
 
   var idSeq = 0;
 
-  var GridStackEngine = function(width, onchange, floatMode, height, items) {
-    this.width = width;
+  var GridStackEngine = function(columns, onchange, floatMode, maxRows, items) {
+    this.columns = columns;
     this.float = floatMode || false;
-    this.height = height || 0;
+    this.maxRows = maxRows || 0;
 
     this.nodes = items || [];
     this.onchange = onchange || function() {};
@@ -320,7 +336,7 @@
     var nn = node;
     var hasLocked = Boolean(this.nodes.find(function(n) { return n.locked; }));
     if (!this.float && !hasLocked) {
-      nn = {x: 0, y: node.y, width: this.width, height: node.height};
+      nn = {x: 0, y: node.y, width: this.columns, height: node.height};
     }
     while (true) {
       var collisionNode = this.nodes.find(Utils._collisionNodeCheck, {node: node, nn: nn});
@@ -339,7 +355,7 @@
   };
 
   GridStackEngine.prototype._sortNodes = function(dir) {
-    this.nodes = Utils.sort(this.nodes, dir, this.width);
+    this.nodes = Utils.sort(this.nodes, dir, this.columns);
   };
 
   GridStackEngine.prototype._packNodes = function() {
@@ -393,21 +409,31 @@
   GridStackEngine.prototype._prepareNode = function(node, resizing) {
     node = node || {};
     // if we're missing position, have the grid position us automatically (before we set them to 0,0)
-    if (node.x === undefined || node.y === undefined) {
+    if (node.x === undefined || node.y === undefined || node.x === null || node.y === null) {
       node.autoPosition = true;
     }
-    node = Utils.defaults(node, {width: 1, height: 1, x: 0, y: 0});
 
-    node.x = parseInt('' + node.x);
-    node.y = parseInt('' + node.y);
-    node.width = parseInt('' + node.width);
-    node.height = parseInt('' + node.height);
+    // assign defaults for missing required fields
+    var defaults = {width: 1, height: 1, x: 0, y: 0};
+    node = Utils.defaults(node, defaults);
+
+    // convert any strings over
+    node.x = parseInt(node.x);
+    node.y = parseInt(node.y);
+    node.width = parseInt(node.width);
+    node.height = parseInt(node.height);
     node.autoPosition = node.autoPosition || false;
     node.noResize = node.noResize || false;
     node.noMove = node.noMove || false;
 
-    if (node.width > this.width) {
-      node.width = this.width;
+    // check for NaN (in case messed up strings were passed. can't do parseInt() || defaults.x above as 0 is valid #)
+    if (Number.isNaN(node.x))      { node.x = defaults.x; node.autoPosition = true; }
+    if (Number.isNaN(node.y))      { node.y = defaults.y; node.autoPosition = true; }
+    if (Number.isNaN(node.width))  { node.width = defaults.width; }
+    if (Number.isNaN(node.height)) { node.height = defaults.height; }
+
+    if (node.width > this.columns) {
+      node.width = this.columns;
     } else if (node.width < 1) {
       node.width = 1;
     }
@@ -420,11 +446,11 @@
       node.x = 0;
     }
 
-    if (node.x + node.width > this.width) {
+    if (node.x + node.width > this.columns) {
       if (resizing) {
-        node.width = this.width - node.x;
+        node.width = this.columns - node.x;
       } else {
-        node.x = this.width - node.width;
+        node.x = this.columns - node.width;
       }
     }
 
@@ -472,9 +498,9 @@
       this._sortNodes();
 
       for (var i = 0;; ++i) {
-        var x = i % this.width;
-        var y = Math.floor(i / this.width);
-        if (x + node.width > this.width) {
+        var x = i % this.columns;
+        var y = Math.floor(i / this.columns);
+        if (x + node.width > this.columns) {
           continue;
         }
         if (!this.nodes.find(Utils._isAddNodeIntercepted, {x: x, y: y, node: node})) {
@@ -511,13 +537,13 @@
     }
     var hasLocked = Boolean(this.nodes.find(function(n) { return n.locked; }));
 
-    if (!this.height && !hasLocked) {
+    if (!this.maxRows && !hasLocked) {
       return true;
     }
 
     var clonedNode;
     var clone = new GridStackEngine(
-      this.width,
+      this.columns,
       null,
       this.float,
       0,
@@ -540,26 +566,26 @@
         return n !== clonedNode && Boolean(n.locked) && Boolean(n._dirty);
       }));
     }
-    if (this.height) {
-      res &= clone.getGridHeight() <= this.height;
+    if (this.maxRows) {
+      res &= clone.getGridHeight() <= this.maxRows;
     }
 
     return res;
   };
 
   GridStackEngine.prototype.canBePlacedWithRespectToHeight = function(node) {
-    if (!this.height) {
+    if (!this.maxRows) {
       return true;
     }
 
     var clone = new GridStackEngine(
-      this.width,
+      this.columns,
       null,
       this.float,
       0,
       this.nodes.map(function(n) { return $.extend({}, n); }));
     clone.addNode(node);
-    return clone.getGridHeight() <= this.height;
+    return clone.getGridHeight() <= this.maxRows;
   };
 
   GridStackEngine.prototype.isNodeChangedPosition = function(node, x, y, width, height) {
@@ -650,54 +676,30 @@
     this.container = $(el);
 
     /*eslint-disable camelcase */
-    if (opts.handle_class !== undefined) {
-      opts.handleClass = opts.handle_class;
-      obsoleteOpts('handle_class', 'handleClass');
-    }
-    if (opts.item_class !== undefined) {
-      opts.itemClass = opts.item_class;
-      obsoleteOpts('item_class', 'itemClass');
-    }
-    if (opts.placeholder_class !== undefined) {
-      opts.placeholderClass = opts.placeholder_class;
-      obsoleteOpts('placeholder_class', 'placeholderClass');
-    }
-    if (opts.placeholder_text !== undefined) {
-      opts.placeholderText = opts.placeholder_text;
-      obsoleteOpts('placeholder_text', 'placeholderText');
-    }
-    if (opts.cell_height !== undefined) {
-      opts.cellHeight = opts.cell_height;
-      obsoleteOpts('cell_height', 'cellHeight');
-    }
-    if (opts.vertical_margin !== undefined) {
-      opts.verticalMargin = opts.vertical_margin;
-      obsoleteOpts('vertical_margin', 'verticalMargin');
-    }
-    if (opts.min_width !== undefined) {
-      opts.minWidth = opts.min_width;
-      obsoleteOpts('min_width', 'minWidth');
-    }
-    if (opts.static_grid !== undefined) {
-      opts.staticGrid = opts.static_grid;
-      obsoleteOpts('static_grid', 'staticGrid');
-    }
-    if (opts.is_nested !== undefined) {
-      opts.isNested = opts.is_nested;
-      obsoleteOpts('is_nested', 'isNested');
-    }
-    if (opts.always_show_resize_handle !== undefined) {
-      opts.alwaysShowResizeHandle = opts.always_show_resize_handle;
-      obsoleteOpts('always_show_resize_handle', 'alwaysShowResizeHandle');
-    }
+    obsoleteOpts(opts, 'handle_class', 'handleClass');
+    obsoleteOpts(opts, 'item_class', 'itemClass');
+    obsoleteOpts(opts, 'placeholder_class', 'placeholderClass');
+    obsoleteOpts(opts, 'placeholder_text', 'placeholderText');
+    obsoleteOpts(opts, 'cell_height', 'cellHeight');
+    obsoleteOpts(opts, 'vertical_margin', 'verticalMargin');
+    obsoleteOpts(opts, 'min_width', 'minWidth');
+    obsoleteOpts(opts, 'static_grid', 'staticGrid');
+    obsoleteOpts(opts, 'is_nested', 'isNested');
+    obsoleteOpts(opts, 'always_show_resize_handle', 'alwaysShowResizeHandle');
+    obsoleteOpts(opts, 'width', 'columns');
+    obsoleteOpts(opts, 'height', 'maxRows');
+
+    // container attributes
+    obsoleteAttr(this.container, 'data-gs-width', 'data-gs-columns');
+    obsoleteAttr(this.container, 'data-gs-height', 'data-gs-max-rows');
     /*eslint-enable camelcase */
 
     opts.itemClass = opts.itemClass || 'grid-stack-item';
     var isNested = this.container.closest('.' + opts.itemClass).length > 0;
 
     this.opts = Utils.defaults(opts || {}, {
-      width: parseInt(this.container.attr('data-gs-width')) || 12,
-      height: parseInt(this.container.attr('data-gs-height')) || 0,
+      columns: parseInt(this.container.attr('data-gs-columns')) || 12,
+      maxRows: parseInt(this.container.attr('data-gs-max-rows')) || 0,
       itemClass: 'grid-stack-item',
       placeholderClass: 'grid-stack-placeholder',
       placeholderText: '',
@@ -774,7 +776,7 @@
 
     this._initStyles();
 
-    this.grid = new GridStackEngine(this.opts.width, function(nodes, detachNode) {
+    this.grid = new GridStackEngine(this.opts.columns, function(nodes, detachNode) {
       detachNode = detachNode === undefined ? true : detachNode;
       var maxHeight = 0;
       this.nodes.forEach(function(n) {
@@ -794,7 +796,7 @@
         }
       });
       self._updateStyles(maxHeight + 10);
-    }, this.opts.float, this.opts.height);
+    }, this.opts.float, this.opts.maxRows);
 
     if (this.opts.auto) {
       var elements = [];
@@ -806,7 +808,7 @@
             el: el,
             // if x,y are missing (autoPosition) add them to end of list - keep their respective DOM order
             i: (parseInt(el.attr('data-gs-x')) || 100) +
-              (parseInt(el.attr('data-gs-y')) || 100) * _this.opts.width
+              (parseInt(el.attr('data-gs-y')) || 100) * _this.opts.columns
           });
         });
       Utils.sortBy(elements, function(x) { return x.i; }).forEach(function(item) {
@@ -1708,7 +1710,7 @@
 
     var heightData = Utils.parseHeight(val);
 
-    if (this.opts.verticalMarginUnit === heightData.unit && this.opts.height === heightData.height) {
+    if (this.opts.verticalMarginUnit === heightData.unit && this.opts.maxRows === heightData.height) {
       return ;
     }
     this.opts.verticalMarginUnit = heightData.unit;
@@ -1747,7 +1749,8 @@
   };
 
   GridStack.prototype.cellWidth = function() {
-    return Math.round(this.container.outerWidth() / this.opts.width);
+    // TODO: take margin into account ($horizontal_padding in .scss) to make cellHeight='auto' square ? (see 810-many-columns.html)
+    return Math.round(this.container.outerWidth() / this.opts.columns);
   };
 
   GridStack.prototype.getCellFromPixel = function(position, useOffset) {
@@ -1756,7 +1759,7 @@
     var relativeLeft = position.left - containerPos.left;
     var relativeTop = position.top - containerPos.top;
 
-    var columnWidth = Math.floor(this.container.width() / this.opts.width);
+    var columnWidth = Math.floor(this.container.width() / this.opts.columns);
     var rowHeight = Math.floor(this.container.height() / parseInt(this.container.attr('data-gs-current-height')));
 
     return {x: Math.floor(relativeLeft / columnWidth), y: Math.floor(relativeTop / rowHeight)};
@@ -1805,11 +1808,11 @@
   };
 
   GridStack.prototype.setGridWidth = function(gridWidth,doNotPropagate) {
-    this.container.removeClass('grid-stack-' + this.opts.width);
+    this.container.removeClass('grid-stack-' + this.opts.columns);
     if (doNotPropagate !== true) {
-      this._updateNodeWidths(this.opts.width, gridWidth);
+      this._updateNodeWidths(this.opts.columns, gridWidth);
     }
-    this.opts.width = gridWidth;
+    this.opts.columns = gridWidth;
     this.grid.width = gridWidth;
     this.container.addClass('grid-stack-' + gridWidth);
   };